cl-macosx-minver.patch
See 66d3b87267a52a389fd9a22cf823224cbee8862d
    689f8248fe85bfc72f8f631d8bfce83df82eb271
    4d0dd6cdf121830e6456f3ae327082902f5581a4
Add minver to Cryptlib build as well.
Bump the default min supported mac version to 12.0

For Synchronet "stuff", we can trivially change this, but some work
needs to be done to enable Spidermonkey and Cryptlib to pick up the
chosen value.

JS should be fairly simple since it's a configure argument.  Cryptlib
is a bit trickier since it's in a patch.

Default to building for the current macOS version, fix 3rdp

Now if you don't specify MIN_MAC_OSX_VERSION, it will use the
"current" version of macOS as the target.  "Current" is defined as
X.0 where the major version (X) is greater than 10 and X.Y when
major version is 10.

This allows building for a specific target, but uses the newest
libraries if you don't specify.