cl-read-timeout-every-time.patch
See ac6a78f964edf3f99c102981f58c9e54f6f29152
    8c9b3a0ecbec900702441b4ab2c5fce7e35d2dc4
Proper fix for wiki edit failures.

The patch that sets the read timeout every time for getData() was
flawed, because getData() itself will sometimes set the read timeout
for the next call.

The patch was intended to ensure that the write timeout never
clobbers the read timeout, so now we track if getData() explicitly
set the timeout and if it did, we don't reset it.