cl-allow-none-auth-svr.patch
See 5f45188030b7eb31806b409683b6cebc194cc5e0
Add a patch to allow the server to accept the "none" auth method.

If the SSH_ANYAUTH option is set, the server will accept the "none"
authentication method, and not even suggest the client send a
password or public key.

The client must still send a user ID, but we just ignore it
completely and don't even log it.